+++ packages/legend-studio-app/src/index.tsx
@@ -17,44 +17,70 @@
 import {
   LegendStudio,
   DSLExternalFormat_LegendStudioApplicationPlugin,
-  type LegendStudioApplicationPlugin,
 } from '@finos/legend-studio';
-import { type AbstractPreset, WebConsole } from '@finos/legend-shared';
-import { getLegendGraphExtensionCollection } from '@finos/legend-graph-extension-collection';
-import { QueryBuilder_LegendStudioApplicationPreset } from '@finos/legend-studio-extension-query-builder';
-import { DSLText_LegendStudioApplicationPlugin } from '@finos/legend-extension-dsl-text';
-import { DSLDiagram_LegendStudioApplicationPlugin } from '@finos/legend-extension-dsl-diagram';
-import { DSLDataSpace_LegendStudioApplicationPlugin } from '@finos/legend-extension-dsl-data-space';
-import { DSLPersistence_LegendStudioApplicationPlugin } from '@finos/legend-extension-dsl-persistence';
-import { ESService_LegendStudioApplicationPlugin } from '@finos/legend-extension-external-store-service';
+import {
+  type AbstractPreset,
+  type AbstractPlugin,
+  WebConsole,
+} from '@finos/legend-shared';
+import {
+  DSLText_GraphManagerPreset,
+  DSLText_LegendStudioApplicationPlugin,
+} from '@finos/legend-extension-dsl-text';
+import {
+  DSLDiagram_GraphManagerPreset,
+  DSLDiagram_LegendStudioApplicationPlugin,
+} from '@finos/legend-extension-dsl-diagram';
+import {
+  DSLDataSpace_GraphManagerPreset,
+  DSLDataSpace_LegendStudioApplicationPlugin,
+} from '@finos/legend-extension-dsl-data-space';
+import {
+  DSLPersistence_GraphManagerPreset,
+  DSLPersistence_LegendStudioApplicationPlugin,
+} from '@finos/legend-extension-dsl-persistence';
+import {
+  ESService_GraphManagerPreset,
+  ESService_LegendStudioApplicationPlugin,
+} from '@finos/legend-extension-external-store-service';
 import { ELMorphir_LegendStudioApplicationPlugin } from '@finos/legend-extension-external-language-morphir';
-
-export const getLegendStudioPresetCollection = (): AbstractPreset[] => [
-  ...getLegendGraphExtensionCollection(),
-  new QueryBuilder_LegendStudioApplicationPreset(),
-];
-
-export const getLegendStudioPluginCollection =
-  (): LegendStudioApplicationPlugin[] => [
-    new DSLText_LegendStudioApplicationPlugin(),
-    new DSLDiagram_LegendStudioApplicationPlugin(),
-    new DSLDataSpace_LegendStudioApplicationPlugin(),
-    new DSLExternalFormat_LegendStudioApplicationPlugin(),
-    new DSLPersistence_LegendStudioApplicationPlugin(),
-    new ESService_LegendStudioApplicationPlugin(),
-    new ELMorphir_LegendStudioApplicationPlugin(),
-  ];
+import { DSLExternalFormat_GraphPreset } from '@finos/legend-graph';
+import { DSLPersistenceCloud_GraphManagerPreset } from '@finos/legend-extension-dsl-persistence-cloud';
+import { EFJSONSchema_GraphManagerPreset } from '@finos/legend-extension-external-format-json-schema';
 
 export class LegendStudioWebApplication {
+  static getPresetCollection(): AbstractPreset[] {
+    return [
+      new DSLText_GraphManagerPreset(),
+      new DSLDiagram_GraphManagerPreset(),
+      new DSLDataSpace_GraphManagerPreset(),
+      new DSLExternalFormat_GraphPreset(),
+      new DSLPersistence_GraphManagerPreset(),
+      new DSLPersistenceCloud_GraphManagerPreset(),
+      new EFJSONSchema_GraphManagerPreset(),
+      new ESService_GraphManagerPreset(),
+    ];
+  }
+
+  static getPluginCollection(): AbstractPlugin[] {
+    return [
+      new DSLText_LegendStudioApplicationPlugin(),
+      new DSLDiagram_LegendStudioApplicationPlugin(),
+      new DSLDataSpace_LegendStudioApplicationPlugin(),
+      new DSLExternalFormat_LegendStudioApplicationPlugin(),
+      new DSLPersistence_LegendStudioApplicationPlugin(),
+      new ESService_LegendStudioApplicationPlugin(),
+      new ELMorphir_LegendStudioApplicationPlugin(),
+      // loggers
+      new WebConsole(),
+    ];
+  }
+
   static run(baseUrl: string): void {
     LegendStudio.create()
       .setup({ baseUrl })
-      .withPresets([...getLegendStudioPresetCollection()])
-      .withPlugins([
-        ...getLegendStudioPluginCollection(),
-        // loggers
-        new WebConsole(),
-      ])
+      .withPresets(LegendStudioWebApplication.getPresetCollection())
+      .withPlugins(LegendStudioWebApplication.getPluginCollection())
       .start()
       .catch((e: unknown) => {
         throw e;