+++ packages/legend-studio-app/CHANGELOG.md
@@ -132,7 +132,7 @@
 
 ### Patch Changes
 
-- [#692](https://github.com/finos/legend-studio/pull/692) [`caab0e67`](https://github.com/finos/legend-studio/commit/caab0e6772181e514b246fe6030a02e7169952cc) ([@akphi](https://github.com/akphi)) - Use `@finos/legend-graph-extension-collection` to manage of graph presets.
+- [#692](https://github.com/finos/legend-studio/pull/692) [`caab0e67`](https://github.com/finos/legend-studio/commit/caab0e6772181e514b246fe6030a02e7169952cc) ([@akphi](https://github.com/akphi)) - Use `@finos/legend-extension-collection` to manage of graph presets.
 
 ## 1.3.0